requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-redis>=5.4.0
psycopg2-binary>=2.9.9
python-json-logger>=2.0.7
elastic-apm>=6.19.0
